+++ NEWS.md
@@ -6,6 +6,7 @@ version 0.3.8
 ## enhancements
 
 * improved speed for `spread` under most conditions, especially larger events, possibly up to 2x.
+* `splitRaster()` can now specify file extension, instead of defaulting to `.grd`
 
 version 0.3.6
 =============